backend/requirements.txt
Flask
flask-cors
requests
python-dotenv
transformers
